960: Allow Multi-Valued Conditions in IAM Policies
The code here assumed that any condition value must be a string, but
condition values may also be arrays. The regex tests succeed in many
cases because the array stringifies to a comma-delimited list of its
elements, but this _also_ means that only the first element is
checked, effectively. When the array-of-arrays is passed back to, say,
`sqsCrossAccount`, the `includes` test _works_, but not in the way
which the author intended. This causes false negatives.

Resolves: #960